backend/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
